AURORA: For GFF4 reference deduplication, use offset and struct index
Instead of just the offset. The field offsets are what counts in the
end, and those are calculate using the struct offset and the struct
template field offsets. This opens up the possibility of running into
collisions, especially with the top level struct.

In fact, collisions happen with the ARL files describing the area
environments in Dragon Age: Origins.