BUILD: Fix adding ".dirty" in CMake builds
Unless I'm mistaken, the '|| echo ".dirty"' seems to make it so that
".dirty" is always attached, because it modifies the exit code.

At least echoing XOREOS_REVDIRT right after the execute_process
suggests that. With the '|| echo ".dirty"', it's always 128 for me, no
matter if the tree is dirty or not. Without, it evaluates to either 0
or 1, correctly showing the dirty-state.