NWSCRIPT: Don't error out when seeing tail recursion in STORESTATEs
Since the stack of a STORESTATE subroutine is thrown away when it
returns, we don't need to care about the state of its stack at the
end. This means we don't need to error out when we see such a
subroutine recurse as its last opcode before the RETN, and can
continue analyzing the script.

For my NWN2 test corpus, this reduces the number of recursion-fails
from 933 to 493 scripts, out of 8535 scripts in total.