NWSCRIPT: Let kTypeResource take precedence over kTypeString
Internally, a resource is a string, and some opcodes operating on
strings are used to operate on resources. For example, EQSS/NEQSS
instructions are used to compare a resource with a string.

Therefore, we now assume that opcodes operating on both a resource and
a string variable as if they're the same type operate on two resource
variables. In effect, we let the resource type override a string type.