NWSCRIPT: Mark logically dead block edges
The original BioWare NWScript compiler had a bug when generating
instructions for short-circuiting if (x || y) conditionals. Instead of
doubling the top stack elements and JMPing to the LOGOR, it does a JZ
to the LOGOR. This would smash the stack if it weren't for the fact
that the stack-smashing true edge of this JZ is logically dead, since
we have already established that the top stack elements is not 0.

In effect, this bug disables the short-circuiting: both x and y are
always evaluated.