AURORA: Add support for GFF3s in NWN premium modules
They're missing the 8-byte Aurora header, and the offsets in the GFF
header are artifically increased by 0x00-0xFF (the first byte of the
premium module's NWM file's MD5).

Since the first offset *should* always be a fixed, known value, we can
calculate the value the offsets need to be correct by, so we don't
need to pass it in.