NWSCRIPT: When looking for a path merge, don't process blocks twice
Processing blocks twice leads to a lot of needless recursion steps and
memory allocation in certain situations. Namely, code that consists of
a lot of "if (foo) break;" will show this behaviour.

One example is en_undrentide.ncs, the OnEnter script of The Lost City
of Undrentide, found in the first Neverwinter Nights expansion, SoU.