GFF2XML: Add an --encoding parameter to override string encodings
Unfortunately, Neverwinter Nights (and maybe other games too?) was
pretty lax when it came to language IDs. The language ID 0 was
commonly not really expected to be in ASCII encoding, but in the
encoding of the game installation. So, for example, the Czech version
of the game read language ID 0 strings in CP-1250.

That lead many non-English community module creators (or translators)
to put their non-English language strings into language ID 0. This
worked for their installation of the game. Unfortunately, this then
made the module not work on other language installations of the game,
but since people with those usually didn't play these modules, it
was seldom discovered (and so really didn't matter).

Since we want to operate on *all* files, no matter the language and
encoding, we now provide an --encoding parameter to override the
encoding in cases where we know the correct encoding, not matter what
they *should* be.