COMMON: Add a new CLI class
The goal of this new class is to simplifie the arguments parsing,
I've take inspiration from glib Commandline option parser:
https://developer.gnome.org/glib/stable/glib-Commandline-option-parser.html,
exept that with glib you need to specifie with a flag the type of each options,
here options type are automatically determine when adding an option
with parser.addOption.

Basically most of the job is made in the parser contructor, here you can
set the name of the program, summary, and arguments that are not options.

Option like "--nwm" are set with addOption function.
"--help" is automatically generated with arguments you've use in
addOption and in the contructor, "--version" is generated too, and
call Version::printVersion.