XML2SSF: Use CLI class
In this file, "Usage: xml2ssf [<options>] [<input file>] <output file>"
change to
"Usage: ./bin/xml2ssf [<options>] <[input file] <output file>>".

I had to use an array of stings to get input file and output file,
this is because input file is the optinal argument and is before ouput file,
I would be harder to handle this case automatically in cli.h, as I would need to
check if there is non optinal argument after the optinal one.

I could do that, but as this file is the only one that need that, I don't think
it worth the effort so I just get all arguments in a vector,
and let some bytes of the old parser do what's left of the job