Fix various issues that prevented xmlfix.cpp from a successful compile with a couple of warnings
In the xmlfix.h file, the fixXMLTag and parseLine functions are public so they can be called from main(). 
In xmlfix.cpp, the scope is added to various calls and the text processing logic is updated to work with Common::UString.
Some limiting formatting changes bring it more in line with the standard.