NWN2: Fork main() code from src/aurora/xmlfix.cpp
This code will be used to create a command tool called fix2xml that will
use the src/aurora/xmlfix.cpp class to filter a NWN2-format XML input
file and output a standard XML format file. It is primarily for testing
against stock and mod XML files for NWN2.